// libs/types/customJwtPayload.ts
import { JwtPayload } from 'jwt-decode';

export interface CustomJwtPayload extends JwtPayload {
	_id: string;
	memberType: string;
	memberStatus: string;
	memberAuthType: string;
	memberPhone: string;
	memberNick: string;
	memberFirstName?: string;
	memberLastName?: string;
	memberEmail: string;
	memberImage?: string;
	memberAddress?: string;
	memberDesc?: string;
	memberFacilities: number;
	memberRank: number;
	memberArticles: number;
	memberPoints: number;
	memberLikes: number;
	memberViews: number;
	memberWarnings: number;
	memberBlocks: number;
}